crud operations

+++ src/main/java/com/bpi/demo/service/MovieService.java
@@ -1,57 +1,54 @@
 package com.bpi.demo.service;
 
 import java.util.List;
-
 import org.springframework.stereotype.Service;
-
 import com.bpi.demo.model.Movie;
 import com.bpi.demo.repository.MovieRepository;
-
 
 @Service
 public class MovieService {
 
 	private final MovieRepository movieRepo;
-	
+
 	public MovieService(MovieRepository movieRepo) {
 		this.movieRepo = movieRepo;
 	}
-	
-	//ADD MOVIE
+
+	// ADD MOVIE
 	public Movie addMovie(Movie movie) {
-		Movie newMovie = new Movie();
-		newMovie.setTitle(movie.getTitle());
-		newMovie.setDirector(movie.getDirector());
-		newMovie.setCastMembers(movie.getCastMembers());
-		newMovie.setReleaseDate(movie.getReleaseDate());
-		return movieRepo.save(newMovie);
+		return movieRepo.save(movie);
 	}
-	
-	//GET MOVIE
+
+	// GET MOVIE
 	public List<Movie> getAllMovies() {
 		return movieRepo.findAll();
 	}
-	
-	//UPDATE MOVIE
+
+	// UPDATE MOVIE
 	public Movie updateMovie(Long id, Movie update) {
 		Movie uMovie = movieRepo.findById(id)
 				.orElseThrow(() -> new RuntimeException("Movie not found."));
-		if(update.getTitle() != null) {
+		if (update.getTitle() != null) {
 			uMovie.setTitle(update.getTitle());
-		}		
-		if(update.getDirector() != null) {
+		}
+		if (update.getDirector() != null) {
 			uMovie.setDirector(update.getDirector());
 		}
-		if(update.getCastMembers() != null) {
+		if (update.getCastMembers() != null) {
 			uMovie.setCastMembers(update.getCastMembers());
 		}
-		if(update.getReleaseDate() != null) {
+		if (update.getReleaseDate() != null) {
 			uMovie.setReleaseDate(update.getReleaseDate());
 		}
 		return movieRepo.save(uMovie);
 	}
-	//DELETE MOVIE
+
+	// DELETE MOVIE
 	public void deleteMovie(Long id) {
+		if (!movieRepo.existsById(id)) {
+			throw new IllegalArgumentException("Movie ID not found.");
+		}
+
 		movieRepo.deleteById(id);
 	}
 }
